mito-ai: extract more shared utilities
