+++ mito-ai/mito_ai/tests/db/connections_test.py
@@ -2,7 +2,8 @@
 # Distributed under the terms of the GNU Affero General Public License v3.0 License.
 
 import requests
-from mito_ai.tests.db.test_db_constants import TOKEN, CONNECTION_JSON
+from mito_ai.tests.db.test_db_constants import CONNECTION_JSON
+from mito_ai.tests.conftest import TOKEN
 
 
 # --- ADD CONNECTION ---
